icons in links and force new tab
display the icons with the website's favicon as pulled form google, and all links clicked open up in a new tab instead of opening in the current new tab start page.   I do not have the options display set up so that you can enable this, i don't know how to implement this, my knowledge of HTML and Java script is not that extensive, sorry.  Maybe some time later on.